Course Ownership

+++ controlers/courses.js
@@ -3,6 +3,7 @@ const Bootcamp = require('../model/Bootcamp')
 const Course = require('../model/Course')
 //======= Libs =======\\
 const mongoose = require('mongoose')
+const ErrorResponse = require('../utils/errorResponse')
 const qs = require('qs')
 // Middlewares
 const asyncHandler = require('../middleware/async')
@@ -82,23 +83,34 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 exports.createCourse = asyncHandler(async (req, res, next) => {
   // get the bootcamp id
   console.log('course.js//createCourse() ')
-  const id = req.params.bootcampId
+  const bootcampId = req.params.bootcampId
+  const userId = req.user.id
+
   // add the id to the body manualy hence the Course schema contaions bootcampID
-  req.body.bootcamp = id
+  req.body.bootcamp = bootcampId
+  req.body.user = userId
 
   let query
-  query = Bootcamp.findById(id)
+  query = Bootcamp.findById(bootcampId)
 
   const bootcamp = await query
 
   if (!bootcamp) {
     return next(
+      new ErrorResponse(`Bootcamp Not found with the id of ${bootcampId}`, 404)
+    )
+  }
+
+  const owner = bootcamp.user.toString()
+  if (owner !== req.user.id && req.user.role !== 'admin') {
+    return next(
       new ErrorResponse(
-        `Bootcamp Not found with the id of ${req.params.id}`,
-        404
+        `Not Authorized to create Course for this bootcamp ${req.user.id}`,
+        401
       )
     )
   }
+
   // we found the bootcamp, create the course
   const course = await Course.create(req.body)
 
@@ -112,7 +124,29 @@ exports.createCourse = asyncHandler(async (req, res, next) => {
 // @route   PUT /api/v1/courses/:id
 // @acsess  Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
+  // 6835f178610ec69b152ccb2b
   const id = req.params.id
+  const course = await Course.findById(id)
+
+  if (!course) {
+    return next(
+      new ErrorResponse(`Course Not found with the id of ${req.params.id}`, 404)
+    )
+  }
+
+  // TODO: move to middleware
+  const looged_user_id = req.user.id
+  const creator_user_id = course.user.toString()
+
+  if (creator_user_id !== looged_user_id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `Not Authorized to update bootcamp for the current user ${req.user.id}`,
+        401
+      )
+    )
+  }
+  // -----
 
   const coursePromise = await Course.findByIdAndUpdate(
     id,
@@ -132,7 +166,7 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
   res.status(200).json({
     succsess: true,
     data: coursePromise,
-    message: 'course updated successfully!',
+    message: 'successfully updated!',
   })
 })
 
@@ -141,18 +175,31 @@ exports.updateCourse = asyncHandler(async (req, res, next) => {
 // @acsess  Private
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
   const id = req.params.id
+  const course = await Course.findById(id)
 
-  const coursePromise = await Course.findByIdAndDelete(id)
-  //   {new:true} : will return the new updated Model
-  if (!coursePromise) {
+  if (!course) {
     return next(
       new ErrorResponse(`Course Not found with the id of ${req.params.id}`, 404)
     )
   }
 
+  // TODO: move to middleware
+  const looged_user_id = req.user.id
+  const creator_user_id = course.user.toString()
+
+  if (creator_user_id !== looged_user_id && req.user.role !== 'admin') {
+    return next(
+      new ErrorResponse(
+        `Not Authorized to delete this Course for the current user ${req.user.id}`,
+        401
+      )
+    )
+  }
+
+  await Course.findByIdAndDelete(id)
+
   res.status(200).json({
     succsess: true,
-    data: coursePromise,
-    message: 'course deleted successfully!',
+    message: 'successfully deleted!',
   })
 })
